requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
google-api-python-client>=2.108.0
scikit-learn>=1.3.2
schedule>=1.2.1
